src_docker/matrix/requirements.txt
boto3
Cython
google-api-core
google-auth
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-resumable-media
googleads
googleapis-common-protos
pandas
pre-commit
pyarrow
pyodbc
pydantic
pytest
pytest-cov
python-dotenv
